requirements.txt
streamlit
spacy
pdfminer.six
python-docx

# SpaCy model for NER - run: python -m spacy download en_core_web_sm
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

